refactor(rfc0020): PR 6 slice 4 — PR-2 review #6/#7/#9/#10/#11 + episode-routing mixin extraction
Apply the PR-2 review residuals carried over into PR 6:

- #6: Rewrite the `_store_event_episode` exception-handler comment to
  enumerate all three failure surfaces (I/O, single-turn close-then-store
  race, tracker programming errors) instead of only narrating the
  store-after-close race.
- #7: Drop the silent `... or interaction` mypy-shim fallback that masked
  an invariant violation; replace with an explicit `RuntimeError` guard
  on the `close()` return value (explicit guard rather than `assert` —
  the latter is stripped under `python -O` per slice-3 precedent).
- #9: Add a runtime telemetry probe to the parity suite — drives a
  `TASK_ASSIGNED` end-to-end through an `InMemoryMetricReader` and
  asserts `agent.interactions.opened` / `closed` / `closed.by_structural`
  each increment once. Closes the gap where a regression dropping the
  close call would have silently zeroed the counters in production.
- #10: Parametrise routing assertions over every member of
  `_SINGLE_TURN_EVENT_TYPES` (TASK_ASSIGNED, SUB_AGENT_COMPLETED,
  APPROVAL_REQUESTED, APPROVAL_RESPONSE, AGENT_JOINED, AGENT_LEFT) so a
  future single-turn type cannot regress the scope-labelling contract.
  Also add explicit coverage for the unknown-event fallback branch
  (warn + legacy NULL-interaction shape).
- #11: Move `AgentAction` / `AgentEvent` under `if TYPE_CHECKING:` —
  they are referenced only in annotations and `from __future__ import
  annotations` is already in effect.

PR-2 review #8 (default `idle_timeout_sec` on `InteractionTracker`) was
already discharged by PR 3's `interaction_idle_timeout_sec` config
plumbing — no further work needed.

File-size split (mechanical): the inline #6/#7/#11 fixes pushed
`state_persistence.py` to 528 lines, over the 500-line cap enforced by
`scripts/checks/file_size.py --strict`. Extract `_store_event_episode`
plus the multi-turn aggregation helpers, the close-path orchestrator,
and the janitor entry point into a new
`agents/persona_runtime/episode_routing.py` (`_EpisodeRoutingMixin`).
`_LLMPersonaAgent` gains the new mixin alongside `_StatePersistenceMixin`;
`close_memory` still calls `drain_pending_summaries` (now provided by
`_EpisodeRoutingMixin`) under `self._lock`. The frozensets move with
the methods that read them; `test_channel_message_runtime.py` updated
to import `_EpisodeRoutingMixin` from the new home.

Mirror split on the test side: PR-2 review #9 + #10 follow-ups land in
`test_interaction_single_turn_parity_followups.py`, with shared
persona-config / mock-LLM / episode-probe / counter-probe helpers
extracted to `_persona_parity_helpers.py`. Both test files now sit
under the file-size cap.

No behaviour change. All 2008 python tests pass (excluding the pre-existing
pattern-parity failure on `main`).
