+++ FILEMAP.md
@@ -1,7 +1,7 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-05-08  
-> **755** tracked files
+> **758** tracked files
 
 ## Summary
 
@@ -9,8 +9,8 @@
 |---------------------|-------|
 | `internal/` | 209 |
 | `docs/` | 155 |
-| `tests/` | 153 |
-| `agents/` | 112 |
+| `tests/` | 155 |
+| `agents/` | 113 |
 | `prompts/` | 24 |
 | `(root)` | 20 |
 | `scripts/` | 20 |
@@ -27,7 +27,7 @@
 
 | Extension | Count |
 |-----------|-------|
-| `.py` | 277 |
+| `.py` | 280 |
 | `.go` | 217 |
 | `.md` | 194 |
 | `.yaml` | 27 |
@@ -154,6 +154,7 @@
 │   │   ├── channel_history.py
 │   │   ├── channel_ingest.py
 │   │   ├── channel_reply.py
+│   │   ├── episode_routing.py
 │   │   ├── memory_budget.py
 │   │   ├── memory_context.py
 │   │   ├── prompt_assembly.py
@@ -733,6 +734,7 @@
 │   ├── integration/
 │   │   ├── __init__.py
 │   │   ├── _delegation_helpers.py
+│   │   ├── _persona_parity_helpers.py
 │   │   ├── _summarize_close_helpers.py
 │   │   ├── _temporal_test_helpers.py
 │   │   ├── audit_logger_integration_test.go
@@ -747,6 +749,7 @@
 │   │   ├── test_delegation_rollback_edges.py
 │   │   ├── test_interaction_multi_turn.py
 │   │   ├── test_interaction_single_turn_parity.py
+│   │   ├── test_interaction_single_turn_parity_followups.py
 │   │   ├── test_log_trace_correlation.py
 │   │   ├── test_logs_correlation.py
 │   │   ├── test_logs_e2e.py
